requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
sqlalchemy==2.0.44
python-jose[cryptography]==3.5.0
pwdlib[argon2,bcrypt]==0.3.0
pydantic==2.12.3
python-dotenv==1.2.1
psycopg2-binary==2.9.11
email-validator==2.3.0
anyio==4.11.0
starlette==0.49.3
watchfiles==1.1.1

